+++ NEWS.md
@@ -7,6 +7,9 @@ when `raw=TRUE`
 
 * Adding `window` option to `pwelch`
 
+* Allowing `downloadEnv`, `matchEnvData`, `erddapToEdinfo`, and `ncToData` to
+work with certain stationary Tabledap datasets (previously only griddap)
+
 # PAMmisc 1.12.3
 
 * Adding `nc='HYCOM'` option to make it easier to specify that you want